+++ udata_piwik/views.py
@@ -1,12 +1,10 @@
 from flask import render_template, Blueprint
-from jinja2 import contextfunction
 
-from udata.frontend import footer_snippet
+from udata.frontend import template_hook
 
 blueprint = Blueprint('piwik', __name__, template_folder='templates')
 
 
-@footer_snippet
-@contextfunction
+@template_hook('footer.snippets')
 def render_piwik_snippet(ctx):
     return render_template('piwik.html', **ctx)